Remove dependence on SPI header
MCP42xx and MCP49xx drivers now take a function pointer to a generic SPI
write function at init, and multiple instances of each peripheral are
also now supported as a result. Added typedefs to represent instances of
the devices themselves as well as configuration settings in the case of
the MCP49xx.